+++ src/components/Inputs/Button/stories/Button.tsx
@@ -1,6 +1,5 @@
-import type { ElementType } from "react";
 import Component, { ButtonProps } from "@/components/Inputs/Button";
 
-const Button = <C extends ElementType>(props: ButtonProps<C>) => <Component {...props} />;
+const Button = (props: ButtonProps) => <Component {...props} />;
 
 export default Button;