+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
 ## v1.39.6
-- **[feat]** - migration to [bun](https://bun.sh/) from ~~[yarn](https://yarnpkg.com/)~~
+- **[chore]** - migration TypeScript to 5 from 4
+- **[chore]** - migration to [bun](https://bun.sh/) from ~~[yarn](https://yarnpkg.com/)~~