+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
-## v1.39.7
-- **[fix]** - props type `Button` 
+## v1.39.8
+- **[fix]** -`NavLink` props  type checking from`NavigationMenu` 
+    component for typescript 5.x.x compliance
+- **[fix]** -`MenuIcon` double semantic `svg` html tag used 
+    in `MobileNavBar` and `TabletNavBar`